+++ audio_service/lib/audio_service.dart
@@ -206,8 +206,17 @@ class PlaybackState {
   /// this calculation is provided by the [position] getter.
   final Duration updatePosition;
 
+  /// for getting stream of current position
+  final Stream<Duration> postionStream;
+
   /// The buffered position.
   final Duration bufferedPosition;
+
+  /// for getting stream of buffered position
+  final Stream<Duration> bufferedStream;
+
+  /// for getting stream of total durtation
+  final Stream<Duration> durationStream;
 
   /// The current playback speed where 1.0 means normal speed.
   final double speed;